feat: add 3D models for footprints
This commit adds 3D models for existing footprints.

Traditionally, 3D models are stored in the `3dmodels` directory and
referenced from corresponding footprints but this is cumbersome because:

- Using `KICAD9_3DMODEL_DIR` is incorrect because the models will never
  be found in the upstream `3dmodels` directory.
- Using `KIPRJMOD` is too prescriptive because it depends on how each
  project imports the project.

Footprints and their 3D model is very closely related to one another so
it may make more sense to embed the 3D model directly into the
footprint, which is something only made possible with KiCad v9.